+++ tests/apps/external_integrations/slack_test.py
@@ -1,8 +1,8 @@
 import pytest
 import responses
 
-from busy_beaver.apps.external_integrations.oauth_providers.slack import SlackOAuthFlow
-from busy_beaver.apps.external_integrations.workflow import (
+from busy_beaver.apps.oauth_integrations.oauth_providers.slack import SlackOAuthFlow
+from busy_beaver.apps.oauth_integrations.workflow import (
     verify_callback_and_save_tokens_in_database,
 )
 from busy_beaver.models import SlackInstallation
