+++ tests/blueprints/slack/event_subscription_test.py
@@ -2,7 +2,7 @@ import pytest
 import responses
 
 from busy_beaver.adapters.slack import TimezoneInfo
-from busy_beaver.apps.external_integrations.oauth_providers.slack import SlackOAuthFlow
+from busy_beaver.apps.oauth_integrations.oauth_providers.slack import SlackOAuthFlow
 from busy_beaver.models import GitHubSummaryConfiguration, SlackInstallation
 from tests._utilities import FakeSlackClient
 
@@ -96,7 +96,7 @@ def test_slack_onboarding_install(client, session, patch_slack):
     # Arrange
     # Step 1 -- User goes to 3rd party website and authenticates app
     # Step 2 -- Create response to send back during token exchange
-    patched_slack = patch_slack("busy_beaver.apps.external_integrations.workflow")
+    patched_slack = patch_slack("busy_beaver.apps.oauth_integrations.workflow")
     responses.add(
         responses.POST,
         SlackOAuthFlow.TOKEN_URL,
@@ -143,7 +143,7 @@ def test_slack_onboarding_invite_bot_to_channel(
 ):
     """TODO deal with situation where bot is invited to multiple channels"""
     # Arrange
-    patched_slack = patch_slack("busy_beaver.apps.external_integrations.workflow")
+    patched_slack = patch_slack("busy_beaver.apps.oauth_integrations.workflow")
     # Create installation in database
     workspace_id = "TXXXXXXXXX"
     authorizing_user_id = "alysivji"
@@ -194,7 +194,7 @@ def test_slack_onboarding_send_bot_configuration(
         tz="America/Chicago", label="Central Daylight Time", offset=-18000
     )
     patched_slack = patch_slack(
-        "busy_beaver.apps.external_integrations.workflow", timezone_info=tz
+        "busy_beaver.apps.oauth_integrations.workflow", timezone_info=tz
     )
     # Create installation in database
     workspace_id = "TXXXXXXXXX"
